@@ -4,24 +4,24 @@ description: Wassword - Wonderful Password Generator
 version: 1.2.1+8
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  share: ^2.0.1
-  url_launcher: ^6.0.3
-  google_fonts: ^2.0.0
-  mobx: ^2.0.0
+  share: ^2.0.4
+  url_launcher: ^6.0.9
+  google_fonts: ^2.1.0
+  mobx: ^2.0.1
   flutter_mobx: ^2.0.0
-  package_info: ^2.0.0
-  fluentui_system_icons: ^1.1.116
+  package_info: ^2.0.2
+  fluentui_system_icons: ^1.1.134
 
 dev_dependencies:
-  build_runner: ^1.12.0
-  mobx_codegen: ^2.0.0
+  build_runner: ^2.0.5
+  mobx_codegen: ^2.0.1+3
   flutter_launcher_icons: ^0.9.0
-  flutter_native_splash: ^1.1.7+1
+  flutter_native_splash: ^1.1.8+4
   flutter_test:
     sdk: flutter
 
